requirements.txt
flask==2.3.3
google-api-python-client==2.100.0
python-dotenv==1.0.0
requests==2.31.0
